requirements.txt
flask>=2.3.0
werkzeug>=2.3.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.2.0
xgboost>=1.7.0
imbalanced-learn>=0.10.0
joblib>=1.2.0
matplotlib>=3.6.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
openpyxl>=3.0.0
